Add readme link to updated fork
Just want to let you know (if you have missed it), that there is a fork of your project with a lot of improvements.
We gave you credit in the readme. Feel free to merge this PR as people will see it better then.

We are also working on even more bugfixes and improvements. Feel free to have a look:
https://github.com/nims11/IPod-Shuffle-4g/pull/25

Happy coding!